requirements.txt
astroid==2.1.0
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
grpcio==1.18.0
grpcio-tools==1.18.0
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
lxml==4.3.2
mccabe==0.6.1
protobuf==3.6.1
pylint==2.2.2
python-slugify==3.0.0
requests==2.21.0
soupsieve==1.8
text-unidecode==1.2
urllib3==1.24.1
wrapt==1.11.1
